pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = ""
authors = [
    {name = "fddiferd",email = "fddiferd@gmail.com"}
]
readme = "README.md"
requires-python = "=3.11"
dependencies = [
    "snowflake-snowpark-python[pandas] (>=1.28.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
